added init code so modules that export an 'init' function get run at startup.  also added 'defaultState' to subwatch to be checked on init
