+++ run.js
@@ -1,6 +1,6 @@
-const config             = require('./config')
-const irc                = require('irc-upd')
-const { commandHandler } = require('./lib/commands')
+const config   = require('./config')
+const irc      = require('irc-upd')
+const { init } = require('./lib/init')
 
 const bot = new irc.Client(config.server, config.userNick, {
   port: config.port,
@@ -19,29 +19,5 @@ const bot = new irc.Client(config.server, config.userNick, {
   messageSplit: config.messageSplit
 })
 
-// listen for pm
-bot.addListener('pm', function(nick) {
-  bot.say(nick, 'Type !help for a list of commands')
-})
+init(bot)
 
-bot.addListener('notice', function (nick, to, text, message) {
-  if (message.args[1].match(/This nickname is registered/g) !== null) {
-    bot.say('NickServ', `identify ${ config.nickservPassword }`)
-  }
-  if (message.args[1].match(/Password accepted - you are now recognized./g) !== null) {
-    bot.join(bot.opt.channels.join(','))
-  }
-})
-
-bot.addListener('message', function(nick, to, text) {
-  // removes ' ' and converts into array
-  const args = text.split(' ')
-
-  const opts = { nick, to, text }
-  if (args[0].startsWith('!')) commandHandler(args, bot, opts)
-
-})
-
-bot.addListener('error', function(message) {
-  console.error('error: ', message)
-})
